optimize the new setglobal! function (JuliaLang/julia#44753)
* add missing gc root for typed global assignment

The codegen declares this is callee rooted (since we do not usually
alllocate), but we might call jl_isa now, which could. The annotations
on this function previously were incorrect.

* test: make exports of modules check ignore Main

Main exports everything, so it is not useful to include in this test.

Refs: https://github.com/JuliaLang/julia/issues/44740

* codegen: optimize setglobal like assignment

Fixes some atomic ordering issues also where we were ignoring the
argument. Uses the new builtin where we could. But also fixes some cases
where we would allocate the global binding at the wrong time (which is a
globally visible side-effect of compilation), or might throw from inside
codegen accidentally there instead.