Don't rely on implicit binding creation by setglobal
As discussed in [1], the implicit creation of bindings through the
setglobal! intrinsic was accidentally added in 1.9 unintentionally
and will be removed (ideally) or at the very least deprecated in 1.11.

The recommended replacement syntax is `Core.eval(mod, Expr(:global, sym))` to
introduce the binding and `invokelatest(setglobal!, mod, sym, val)` to set
it. The invokelatest is not presently required, but may be required
for https://github.com/JuliaLang/julia/pull/54654, so it's included
in the recommendation.

[1] https://github.com/JuliaLang/julia/issues/54607