Fix tests when JULIA_DEPOT_PATH is set
When JULIA_DEPOT_PATH is set, the stdlib depots might not be in DEPOT_PATH. However, because the test does not inherit the env var and instead sets it to something custom, the child processes will get the default. To make sure the tests work properly, just reach into base and query the right values.